Refactoring of single entry point for all flows
