+++ fashion-tryon-compose/src/commonMain/kotlin/com/aiuta/fashionsdk/tryon/compose/ui/AiutaTryOnFlow.kt
@@ -1,23 +1,13 @@
 package com.aiuta.fashionsdk.tryon.compose.ui
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.backhandler.BackHandler
 import com.aiuta.fashionsdk.analytics.events.AiutaAnalyticsSessionEvent
 import com.aiuta.fashionsdk.configuration.AiutaConfiguration
 import com.aiuta.fashionsdk.configuration.features.models.product.ProductItem
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.analytic.sendSessionEvent
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.controller.composition.LocalController
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.controller.deactivateSelectMode
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.controller.navigateBack
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.navigation.NavigationContainer
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.navigation.NavigationInitialisation
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.navigation.NavigationScreen
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.screens.base.transition.controller.isSharingEnable
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.screens.zoom.ZoomedImageScreen
-import com.aiuta.fashionsdk.tryon.compose.ui.internal.screens.zoom.controller.closeZoomImageScreen
+import com.aiuta.fashionsdk.tryon.compose.ui.internal.navigation.NavigationFlow
+import com.aiuta.fashionsdk.tryon.compose.ui.internal.navigation.defaultStartScreen
 
 /**
  * Entry point for the fashion try-on flow.
@@ -38,46 +28,11 @@ public fun AiutaTryOnFlow(
     aiutaConfiguration: AiutaConfiguration,
     productForGeneration: ProductItem,
 ) {
-    NavigationInitialisation(
+    NavigationFlow(
         modifier = modifier,
         aiutaConfiguration = aiutaConfiguration,
         productItem = productForGeneration,
-    ) {
-        sendSessionEvent(AiutaAnalyticsSessionEvent.FlowType.TRY_ON)
-
-        val controller = LocalController.current
-        val scope = rememberCoroutineScope()
-
-        NavigationContainer(modifier = modifier)
-
-        // Move screen here, because full view should be on the top of navigation
-        with(controller) {
-            if (zoomImageController.isSharingEnable()) {
-                ZoomedImageScreen(
-                    modifier = modifier,
-                    screenState = zoomImageController,
-                )
-            }
-        }
-
-        BackHandler {
-            when {
-                controller.zoomImageController.isSharingEnable() -> {
-                    controller.zoomImageController.closeZoomImageScreen(scope)
-                }
-
-                controller.bottomSheetNavigator.sheetState.isVisible -> {
-                    controller.bottomSheetNavigator.hide()
-                }
-
-                controller.currentScreen.value == NavigationScreen.History -> {
-                    // Use custom, because we need deactivate select changePhotoButtonStyle first
-                    controller.deactivateSelectMode()
-                    controller.navigateBack()
-                }
-
-                else -> controller.navigateBack()
-            }
-        }
-    }
+        startScreen = defaultStartScreen(),
+        flowType = AiutaAnalyticsSessionEvent.FlowType.TRY_ON,
+    )
 }
